Refactor agent setup + move settings script
